+++ @xen-orchestra/fs/src/fs.spec.js
@@ -127,8 +127,8 @@ handlers.forEach(url => {
       })
 
       it('should return correct type of attribute', async () => {
-        if (info.available !== undefined) {
-          expect(typeof info.available).toBe('number')
+        if (info.size !== undefined) {
+          expect(typeof info.size).toBe('number')
         }
         if (info.used !== undefined) {
           expect(typeof info.used).toBe('number')