+++ packages/xo-web/src/common/intl/messages.js
@@ -482,7 +482,7 @@ const messages = {
   remotePath: 'Path',
   remoteState: 'State',
   remoteDevice: 'Device',
-  remoteDisk: 'Disk',
+  remoteDisk: 'Disk (Used / Total)',
   remoteOptions: 'Options',
   remoteShare: 'Share',
   remoteAction: 'Action',